moved entry/App decided agains the root of components and moved it to the root of source
