+++ src/entry/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './App';
+import App from '../App';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
